requirements.txt
amqp==5.2.0
arrow==1.3.0
asgiref==3.7.2
autopep8==2.0.4
beautifulsoup4==4.12.3
billiard==4.2.0
binaryornot==0.4.4
black==23.12.1
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cookiecutter==2.6.0
cookiecutter-django==1.11.9
cors==1.0.1
coverage==7.4.1
cron_descriptor==1.4.0
distlib==0.3.8
Django==4.2.10
django-analytical==3.1.0
django-browser-reload==1.12.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-defender==0.9.7
django-environ==0.11.2
django-factory-boy==1.0.0
django-filter==23.5
django-livereload==1.7
django-pytest==0.2.0
django-request==1.6.3
django-tailwind==3.8.0
django-timezone-field==6.1.0
environ==1.0
factory-boy==3.3.0
Faker==22.2.0
filelock==3.13.1
flower==2.0.1
future==0.18.3
gevent==23.9.1
greenlet==3.0.3
humanize==4.9.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.3
kombu==5.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mirakuru==2.5.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pika==1.3.2
platformdirs==4.1.0
pluggy==1.3.0
port-for==0.7.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg==3.1.17
psycopg2==2.9.9
pycodestyle==2.11.1
pycparser==2.21
Pygments==2.17.2
PySocks==1.7.1
pytest==7.4.4
pytest-factoryboy==2.6.0
pytest-faker==2.0.0
pytest-mock==3.12.0
pytest-rabbitmq==3.0.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-file==1.5.1
rich==13.7.0
setuptools==69.0.3
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
text-unidecode==1.3
tldextract==5.1.1
toml==0.10.2
tornado==6.4
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.13
whitenoise==6.6.0
zope.event==5.0
zope.interface==6.1